+++ src/components/CalendarDescription/CalendarDescription.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { format } from 'date-fns';
 import ruLocale from 'date-fns/locale/ru';
 import { placesTextForms } from '../../utils/Constants';
-import declOfNum from '../../utils/utils';
+import { declOfNum } from '../../utils/utils';
 
 function CalendarDescription({
   isOpen, onClose, onActionClick, currentEvent,
