+++ src/components/PopupLogin/PopupLogin.scss
@@ -1,6 +1,12 @@
 .popup__input-error{
-  padding:0;
+  padding: 0 0 10px;
+  margin: -15px 10px 0 ;
+  font-size: 12px;
+  color: red;
+}
+.popup__action-error{
+  padding: 0;
   margin: 0;
-  font-size: 8px;
+  font-size: 12px;
   color: red;
 }
